+++ src/gui/base/ButtonSize.ts
@@ -1,3 +1,8 @@
+/**
+ * Buttons have an interactive area that is larger than the visual size
+ * this size is too large to fit into some injections, so we have
+ * a compact size for that.
+ */
 export const enum ButtonSize {
 	Normal,
 	Compact,
