+++ src/gui/base/InfoBanner.ts
@@ -34,7 +34,7 @@ export class InfoBanner implements Component<InfoBannerAttrs> {
 	view(vnode: Vnode<InfoBannerAttrs>): Children {
 		const {message, icon, helpLink, buttons, type} = vnode.attrs
 		return m(
-			".info-banner.center-vertically.full-width.border-bottom.pr-s.pl.border-radius .mt-xs" + // keep the distance to the bottom of the banner the same in the case that buttons aren't present
+			".info-banner.center-vertically.full-width.border-bottom.pr-s.pl.border-radius.mt-xs" + // keep the distance to the bottom of the banner the same in the case that buttons aren't present
 			(buttons && buttons.length > 0 ? "" : ".pb-s"),
 			{
 				style: {
