+++ src/mail/view/MailGuiUtils.ts
@@ -9,23 +9,16 @@ import {AllIcons} from "../../gui/base/Icon"
 import {Icons} from "../../gui/base/icons/Icons"
 import type {InlineImages} from "./MailViewer"
 import {isApp, isDesktop} from "../../api/common/Env"
-import {neverNull, ofClass, promiseMap} from "@tutao/tutanota-utils"
+import {neverNull, promiseMap} from "@tutao/tutanota-utils"
 import {MailFolderType, MailReportType} from "../../api/common/TutanotaConstants"
 import {getElementId} from "../../api/common/utils/EntityUtils"
 import {reportMailsAutomatically} from "./MailReportDialog"
 import {DataFile} from "../../api/common/DataFile";
-import {InfoLink, lang, TranslationKey} from "../../misc/LanguageViewModel"
+import {TranslationKey} from "../../misc/LanguageViewModel"
 import {FileController} from "../../file/FileController"
-import {DomRectReadOnlyPolyfilled, Dropdown, DropdownButtonAttrs, PosRect} from "../../gui/base/Dropdown.js"
+import {DomRectReadOnlyPolyfilled, Dropdown, PosRect} from "../../gui/base/Dropdown.js"
 import {ButtonSize} from "../../gui/base/ButtonSize.js"
 import {modal} from "../../gui/base/Modal.js"
-import {ContentBlockingStatus, MailViewerViewModel} from "./MailViewerViewModel.js"
-import m from "mithril"
-import {ifAllowedTutanotaLinks} from "../../gui/base/GuiUtils.js"
-import {Button, ButtonType} from "../../gui/base/Button.js"
-import {client} from "../../misc/ClientDetector.js"
-import {showProgressDialog} from "../../gui/dialogs/ProgressDialog.js"
-import {showHeaderDialog} from "./MailViewerUtils.js"
 
 export function showDeleteConfirmationDialog(mails: ReadonlyArray<Mail>): Promise<boolean> {
 	let groupedMails = mails.reduce(
@@ -314,149 +307,3 @@ export function showMoveMailsDropdown(model: MailModel, origin: PosRect, mails: 
 	})
 }
 
-export function mailViewerMoreActions(
-	viewModel: MailViewerViewModel,
-) {
-	const moreButtons: Array<DropdownButtonAttrs> = []
-	if (viewModel.isUnread()) {
-		moreButtons.push({
-			label: "markRead_action",
-			click: () => viewModel.setUnread(false),
-			icon: Icons.Eye,
-		})
-	} else {
-		moreButtons.push({
-			label: "markUnread_action",
-			click: () => viewModel.setUnread(true),
-			icon: Icons.NoEye,
-		})
-	}
-
-	if (!client.isMobileDevice() && viewModel.canExport()) {
-		moreButtons.push({
-			label: "export_action",
-			click: () => showProgressDialog("pleaseWait_msg", viewModel.exportMail()),
-			icon: Icons.Export,
-		})
-	}
-
-	if (!client.isMobileDevice() && typeof window.print === "function" && viewModel.canPrint()) {
-		moreButtons.push({
-			label: "print_action",
-			click: () => window.print(),
-			icon: Icons.Print,
-		})
-	}
-
-	if (viewModel.isListUnsubscribe()) {
-		moreButtons.push({
-			label: "unsubscribe_action",
-			click: () => unsubscribe(viewModel),
-			icon: Icons.Cancel,
-		})
-	}
-
-	if (viewModel.canShowHeaders()) {
-		moreButtons.push({
-			label: "showHeaders_action",
-			click: () => showHeaderDialog(viewModel.getHeaders()),
-			icon: Icons.ListUnordered,
-		})
-	}
-
-	if (viewModel.canReport()) {
-		moreButtons.push({
-			label: "reportEmail_action",
-			click: () => reportMail(viewModel),
-			icon: Icons.Warning,
-		})
-	}
-
-	if (viewModel.canPersistBlockingStatus() && viewModel.isShowingExternalContent()) {
-		moreButtons.push({
-			label: "disallowExternalContent_action",
-			click: () => viewModel.setContentBlockingStatus(ContentBlockingStatus.Block),
-			icon: Icons.Picture,
-		})
-	}
-
-	if (viewModel.canPersistBlockingStatus() && viewModel.isBlockingExternalImages()) {
-		moreButtons.push({
-			label: "showImages_action",
-			click: () => viewModel.setContentBlockingStatus(ContentBlockingStatus.Show),
-			icon: Icons.Picture,
-		})
-	}
-
-	return moreButtons
-}
-
-function unsubscribe(viewModel: MailViewerViewModel): Promise<void> {
-	return showProgressDialog("pleaseWait_msg", viewModel.unsubscribe())
-		.then(success => {
-			if (success) {
-				return Dialog.message("unsubscribeSuccessful_msg")
-			}
-		})
-		.catch(e => {
-			if (e instanceof LockedError) {
-				return Dialog.message("operationStillActive_msg")
-			} else {
-				return Dialog.message("unsubscribeFailed_msg")
-			}
-		})
-}
-
-function reportMail(viewModel: MailViewerViewModel) {
-	const sendReport = (reportType: MailReportType) => {
-		viewModel.reportMail(reportType)
-				 .catch(ofClass(LockedError, () => Dialog.message("operationStillActive_msg")))
-				 .finally(m.redraw)
-	}
-
-	const dialog = Dialog.showActionDialog({
-		title: lang.get("reportEmail_action"),
-		child: () =>
-			m(
-				".flex.col.mt-m",
-				{
-					// So that space below buttons doesn't look huge
-					style: {
-						marginBottom: "-10px",
-					},
-				},
-				[
-					m("div", lang.get("phishingReport_msg")),
-					ifAllowedTutanotaLinks(InfoLink.Phishing, link =>
-						m(
-							"a.mt-s",
-							{
-								href: link,
-								target: "_blank",
-							},
-							lang.get("whatIsPhishing_msg"),
-						),
-					),
-					m(".flex-wrap.flex-end", [
-						m(Button, {
-							label: "reportPhishing_action",
-							click: () => {
-								sendReport(MailReportType.PHISHING)
-								dialog.close()
-							},
-							type: ButtonType.Secondary,
-						}),
-						m(Button, {
-							label: "reportSpam_action",
-							click: () => {
-								sendReport(MailReportType.SPAM)
-								dialog.close()
-							},
-							type: ButtonType.Secondary,
-						}),
-					]),
-				],
-			),
-		okAction: null,
-	})
-}
