+++ src/misc/TranslationKey.ts
@@ -1477,3 +1477,4 @@ export type TranslationKeyType =
 	| "yourMessage_label"
 	| "you_label"
 	| "emptyString_msg"
+	| "mailViewerRecipients_label"
