+++ src/translations/de.ts
@@ -1494,6 +1494,7 @@ export default {
 		"yourCalendars_label": "Deine Kalender",
 		"yourFolders_action": "DEINE ORDNER",
 		"yourMessage_label": "Deine Nachricht",
-		"you_label": "Du"
+		"you_label": "Du",
+		"mailViewerRecipients_label": "an:",
 	}
 }
