+++ src/translations/de_sie.ts
@@ -1494,6 +1494,7 @@ export default {
 		"yourCalendars_label": "Deine Kalender",
 		"yourFolders_action": "Ihre ORDNER",
 		"yourMessage_label": "Ihre Nachricht",
-		"you_label": "Sie"
+		"you_label": "Sie",
+		"mailViewerRecipients_label": "an:",
 	}
 }
