+++ src/translations/en.ts
@@ -1490,6 +1490,7 @@ export default {
 		"yourCalendars_label": "Your calendars",
 		"yourFolders_action": "YOUR FOLDERS",
 		"yourMessage_label": "Your message",
-		"you_label": "You"
+		"you_label": "You",
+		"mailViewerRecipients_label": "to:",
 	}
 }
